# docs/swagger.yaml
basePath: /api
definitions:
  models.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple user authentication API using Go Gin and MongoDB
  title: Go Gin Mongo API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get All Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get All Users
      tags:
      - User
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get User by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get User by ID
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: Login User
      parameters:
      - description: User Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.User'
      summary: Login User
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Register User
      parameters:
      - description: User Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.User'
      summary: Register User
      tags:
      - User
schemes:
- http
swagger: "2.0"